vibe_coding/grand-things/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-multipart>=0.0.6
jieba>=0.42.0
python-dateutil>=2.8.0
# 用户认证相关依赖
bcrypt>=4.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
cryptography>=41.0.0 